% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datageneration.R
\name{datageneration}
\alias{datageneration}
\title{Generate Artificial Data from Simulation}
\usage{
datageneration(
  n,
  design = "complete",
  dgp = 1,
  beta_d = 1,
  beta_s = 0,
  Y0 = NULL,
  A = "Erdos-Renyi"
)
}
\arguments{
\item{n}{An even number specifying the sample size}

\item{design}{A character specifying the randomization design.
Options are "Bernoulli", "complete", and "stratified",
which stand for Bernoulli randomization, complete randomization, and
stratified randomization.
Default is "complete".}

\item{dgp}{Specify data generating process 1 or 2 for the outcome equation}

\item{beta_d}{A scalar specifying the value
of a coefficient for the direct effect in the outcome equation.
There is no direct effect if beta_d is set to zero.}

\item{beta_s}{A scalar specifying the value
of a coefficient for spillovers in the outcome equation.
There are no spillovers if beta_s is set to zero.}

\item{Y0}{NULL or an n-dimensional vector of the untreated potential outcomes.
Default is NULL.}

\item{A}{An n times n binary adjacency matrix or
a character specifying how to generate the adjacency matrix.
For the latter case, options are "Erdos-Renyi" and "pairs".
Default is "Erdos-Renyi".}
}
\value{
A list containing the following elements.
\item{Y}{An n-dimensional outcome vector}
\item{Z}{An n-dimensional treatment assignment vector}
\item{A}{An n times n binary adjacency matrix}
\item{strata}{An n-dimensional vector indicating strata}
}
\description{
Generate Artificial Data from Simulation
}
\examples{
data <- datageneration(n = 200)

}
